Refactored code: Added Prediction and Scheduler services. Modifed Leaderboard and Player services.
